[marvell][platform] Checking file presence before access (#6208)
In marvell_et644m platform scripts, I have added a check to confirm the file availability before accessing it.

+++ platform/marvell-armhf/sonic-platform-et6448m/entropy.py
@@ -1,6 +1,7 @@
 #!/usr/bin/python
 import fcntl, struct
 import time
+from os import path
 
 RNDADDENTROPY=0x40085203
 
@@ -8,10 +9,11 @@ def avail():
   with open("/proc/sys/kernel/random/entropy_avail", mode='r') as avail:
       return int(avail.read())
 
-while 1:
-    while avail() < 2048:
-        with open('/dev/urandom', 'rb') as urnd, open("/dev/random", mode='wb') as rnd:
-            d = urnd.read(512)
-            t = struct.pack('ii', 4 * len(d), len(d)) + d
-            fcntl.ioctl(rnd, RNDADDENTROPY, t)
-    time.sleep(30)
+if path.exists("/proc/sys/kernel/random/entropy_avail"):
+    while 1:
+        while avail() < 2048:
+            with open('/dev/urandom', 'rb') as urnd, open("/dev/random", mode='wb') as rnd:
+                d = urnd.read(512)
+                t = struct.pack('ii', 4 * len(d), len(d)) + d
+                fcntl.ioctl(rnd, RNDADDENTROPY, t)
+        time.sleep(30)
